Merge davide's branch

Approved-by: Emanuele Concas <conema@unica.it>

+++ CheChatApp/views.py
@@ -101,7 +101,7 @@ def logout(request):
     auth_logout(request)
     return redirect('login')
 
-
+   #tested
 def info_chat_by_id(request, chat_id):
     chat = Chat.objects.filter(id=chat_id)
 
@@ -118,7 +118,7 @@ def info_chat_by_id(request, chat_id):
 
     return JsonResponse(response)
 
-
+    #tested
 def get_chat_by_user(request):
     """Get chat of the user"""
 
@@ -132,7 +132,7 @@ def get_chat_by_user(request):
         }
     else:
         response = {
-            'state': 'chat doesn\'t exists'
+            'state': 'chat doesn\'t exist'
         }
 
     return JsonResponse(response)
@@ -156,17 +156,20 @@ def get_last_message(request, chat_id):
                 'message': []
             }
 
+    elif not is_participants(chat_id, request.user.id):
+        response = {
+            'state': 'not a participant'
+        }
     else:
         response = {
-            'state': 'chat does not exist'
+            'state': 'chat not found'
         }
 
     return JsonResponse(response)
 
-
+    #tested
 def new_chat(request, title=""):
     """Create a new"""
-
     # TODO: controllare se l'utente è amico
     if request.user.is_authenticated:
         chat = Chat.objects.create(title=title)
@@ -235,14 +238,19 @@ def get_participants(request, chat_id):
 
     return JsonResponse(response)
 
-
+    #tested
 def get_contacts(request):
-    phonebook = PhoneBook.objects.filter(owner=request.user)
+    try:
+        phonebook = PhoneBook.objects.get(owner=request.user)
+    except PhoneBook.DoesNotExist:
+        phonebook = None
+        response = {'state': 'phonebook not found'}
 
-    response = {
-        'state': 'successful',
-        'contacts': list(phonebook.values_list('contacts', flat=True))
-    }
+    if phonebook is not None:
+        response = {
+            'state': 'successful',
+            'contacts': list(phonebook.contacts.values_list(flat=True))
+        }
 
     return JsonResponse(response)
 
@@ -265,16 +273,19 @@ def add_contact(request, added_user_id):
 
     return JsonResponse(response)
 
-
+    #tested
 def delete_contact(request, user_to_delete_id):
-    phonebook = PhoneBook.objects.get(owner=request.user)
-    response = {}
+    try:
+        phonebook = PhoneBook.objects.get(owner=request.user)
+    except PhoneBook.DoesNotExist:
+        phonebook = None
+        response = {'state': 'phonebook not found'}
 
-    if User.objects.filter(id=user_to_delete_id).exists() \
+    if User.objects.filter(id=user_to_delete_id).exists() and phonebook is not None \
             and phonebook.contacts.filter(id=user_to_delete_id).exists():
         response = {'state': 'successful'}
         phonebook.contacts.get(id=user_to_delete_id).delete()
-    elif phonebook.contacts.filter(id=user_to_delete_id).exists():
+    elif phonebook is not None and not phonebook.contacts.filter(id=user_to_delete_id).exists():
         response = {'state': 'user does not exist'}
 
     return JsonResponse(response)
@@ -317,7 +328,7 @@ def get_messages_by_id(request, chat_id):
 
     return JsonResponse(response)
 
-
+    #tested
 def is_participants(chat_id, user_id):
     """Check if the user is a participant of the chat"""
